Improve sources files separation and continue to convert SetInputData to SetInputConnection
